refactor: implement new auth to all modules

+++ src/gizi-wise/user/user.controller.ts
@@ -1,25 +1,34 @@
 import {
   Controller,
   Get,
-  Post,
   Body,
   Patch,
   Param,
   Delete,
   Query,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { ApiBody, ApiTags } from '@nestjs/swagger';
+import { ApiTags } from '@nestjs/swagger';
 import { QueryListUserDto } from './dto/query-list-user.dto';
 import { ResponseListUserDto } from './dto/list-user.dto';
+import { Role } from '@gizi-wise/admin/entities/admin.entity';
+import { Auth } from '@gizi-wise/auth/auth.decorator';
+import { LoggedUser } from '@common/decorators/logged-user.decorator';
 
 @ApiTags('Users')
 @Controller('users')
 export class UserController {
+  errorMessages = {
+    forbidSee: 'You are not allowed to see other user info',
+    forbidModify: 'You are not allowed to modify other user',
+    forbidDelete: 'You are not allowed to delete other user',
+  };
   constructor(private readonly userService: UserService) {}
 
   @Get()
+  @Auth(Role.ADMIN)
   async findAll(@Query() query: QueryListUserDto) {
     const { page, limit } = query;
     query.offset = (page - 1) * limit;
@@ -33,32 +42,33 @@ export class UserController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  @Auth(Role.USER)
+  findOne(@LoggedUser() user: any, @Param('id') id: string) {
+    if (user.role === Role.USER && user.id !== id) {
+      throw new ForbiddenException(this.errorMessages.forbidSee);
+    }
     return this.userService.findOne(id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  @Auth(Role.USER)
+  update(
+    @LoggedUser() user: any,
+    @Param('id') id: string,
+    @Body() updateUserDto: UpdateUserDto,
+  ) {
+    if (user.role === Role.USER && user.id !== id) {
+      throw new ForbiddenException(this.errorMessages.forbidModify);
+    }
     return this.userService.update(id, updateUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @Auth(Role.USER)
+  remove(@LoggedUser() user: any, @Param('id') id: string) {
+    if (user.role === Role.USER && user.id !== id) {
+      throw new ForbiddenException(this.errorMessages.forbidDelete);
+    }
     return this.userService.remove(id);
-  }
-
-  @Post('firebase-sign-in')
-  @ApiBody({
-    schema: {
-      type: 'object',
-      properties: {
-        idToken: {
-          type: 'string',
-        },
-      },
-    },
-  })
-  verifyIdToken(@Body('idToken') idToken: string) {
-    return this.userService.verifyIdToken(idToken);
   }
 }
